+++ Editor/Modules/BuildReportModule.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using Unity.ProjectAuditor.Editor.Core;
 using Unity.ProjectAuditor.Editor.Diagnostic;
 using Unity.ProjectAuditor.Editor.Utils;
+using Unity.ProjectAuditor.Editor.Build;
 using UnityEditor;
 using UnityEditor.Build.Reporting;
 using UnityEditor.Callbacks;
@@ -34,63 +35,13 @@ namespace Unity.ProjectAuditor.Editor.Modules
         Num
     }
 
-    public interface IBuildReportProvider
-    {
-        BuildReport GetBuildReport();
-    }
-
-    class LastBuildReportProvider : IBuildReportProvider
-    {
-        internal const string k_LastBuildReportPath = "Library/LastBuild.buildreport";
-
-        public BuildReport GetBuildReport()
-        {
-            return GetLastBuildReportAsset();
-        }
-
-        public static BuildReport GetLastBuildReportAsset()
-        {
-            if (!File.Exists(k_LastBuildReportPath))
-                return null; // a build report was not found in the Library folder
-
-            var buildReportPath = UserPreferences.buildReportPath;
-            if (!Directory.Exists(buildReportPath))
-                Directory.CreateDirectory(buildReportPath);
-
-            var date = File.GetLastWriteTime(k_LastBuildReportPath);
-            var targetAssetName = "Build_" + date.ToString("yyyy-MM-dd-HH-mm-ss");
-            var assetPath = $"{buildReportPath}/{targetAssetName}.buildreport";
-
-            if (!File.Exists(assetPath))
-            {
-                var tempAssetPath = buildReportPath + "/New Report.buildreport";
-                File.Copy(k_LastBuildReportPath, tempAssetPath, true);
-                AssetDatabase.ImportAsset(tempAssetPath);
-                AssetDatabase.RenameAsset(tempAssetPath, targetAssetName);
-            }
-
-            return AssetDatabase.LoadAssetAtPath<BuildReport>(assetPath);
-        }
-
-        [PostProcessBuild(1)]
-        public static void OnPostprocessBuild(BuildTarget target, string pathToBuiltProject)
-        {
-            if (UserPreferences.buildReportAutoSave)
-            {
-                // Library/LastBuild.buildreport is only created AFTER OnPostprocessBuild so we need to defer the copy of the file
-                EditorApplication.update += CheckLastBuildReport;
-            }
-        }
-
-        static void CheckLastBuildReport()
-        {
-            if (GetLastBuildReportAsset() != null)
-                EditorApplication.update -= CheckLastBuildReport;
-        }
-    }
-
     public class BuildReportModule : ProjectAuditorModule
     {
+        internal interface IBuildReportProvider
+        {
+            BuildReport GetBuildReport(BuildTarget platform);
+        }
+
 #if BUILD_REPORT_API_SUPPORT
         const string k_KeyBuildPath = "Path";
         const string k_KeyPlatform = "Platform";
@@ -142,13 +93,13 @@ namespace Unity.ProjectAuditor.Editor.Modules
         static IBuildReportProvider s_BuildReportProvider;
         static IBuildReportProvider s_DefaultBuildReportProvider = new LastBuildReportProvider();
 
-        public static IBuildReportProvider BuildReportProvider
+        internal static IBuildReportProvider BuildReportProvider
         {
             get => s_BuildReportProvider != null ? s_BuildReportProvider : s_DefaultBuildReportProvider;
             set => s_BuildReportProvider = value;
         }
 
-        public static IBuildReportProvider DefaultBuildReportProvider => s_BuildReportProvider;
+        internal static IBuildReportProvider DefaultBuildReportProvider => s_BuildReportProvider;
 
         public override string name => "Build Report";
 
@@ -166,48 +117,46 @@ namespace Unity.ProjectAuditor.Editor.Modules
         public override void Audit(ProjectAuditorParams projectAuditorParams, IProgress progress = null)
         {
 #if BUILD_REPORT_API_SUPPORT
-            var buildReport = BuildReportProvider.GetBuildReport();
+            var buildReport = BuildReportProvider.GetBuildReport(projectAuditorParams.platform);
             if (buildReport != null)
             {
-                var issues = new List<ProjectIssue>();
-                NewMetaData(k_KeyBuildPath, buildReport.summary.outputPath, issues);
-                NewMetaData(k_KeyPlatform, buildReport.summary.platform, issues);
-                NewMetaData(k_KeyResult, buildReport.summary.result, issues);
-                NewMetaData(k_KeyStartTime, Formatting.FormatDateTime(buildReport.summary.buildStartedAt), issues);
-                NewMetaData(k_KeyEndTime, Formatting.FormatDateTime(buildReport.summary.buildEndedAt), issues);
-                NewMetaData(k_KeyTotalTime, Formatting.FormatBuildTime(buildReport.summary.totalTime), issues);
-                NewMetaData(k_KeyTotalSize, Formatting.FormatSize(buildReport.summary.totalSize), issues);
-
-                AnalyzeBuildSteps(buildReport, issues);
-                AnalyzePackedAssets(buildReport, issues);
-
-                if (issues.Any())
-                    projectAuditorParams.onIncomingIssues(issues);
+                projectAuditorParams.onIncomingIssues(new[]
+                {
+                    NewMetaData(k_KeyBuildPath, buildReport.summary.outputPath),
+                    NewMetaData(k_KeyPlatform, buildReport.summary.platform),
+                    NewMetaData(k_KeyResult, buildReport.summary.result),
+                    NewMetaData(k_KeyStartTime, Formatting.FormatDateTime(buildReport.summary.buildStartedAt)),
+                    NewMetaData(k_KeyEndTime, Formatting.FormatDateTime(buildReport.summary.buildEndedAt)),
+                    NewMetaData(k_KeyTotalTime, Formatting.FormatBuildTime(buildReport.summary.totalTime)),
+                    NewMetaData(k_KeyTotalSize, Formatting.FormatSize(buildReport.summary.totalSize)),
+                });
+                projectAuditorParams.onIncomingIssues(AnalyzeBuildSteps(buildReport));
+                projectAuditorParams.onIncomingIssues(AnalyzePackedAssets(buildReport));
             }
 #endif
             projectAuditorParams.onModuleCompleted?.Invoke();
         }
 
 #if BUILD_REPORT_API_SUPPORT
-        void AnalyzeBuildSteps(BuildReport buildReport, IList<ProjectIssue> issues)
+        IEnumerable<ProjectIssue> AnalyzeBuildSteps(BuildReport buildReport)
         {
             foreach (var step in buildReport.steps)
             {
                 var depth = step.depth;
-                issues.Add(ProjectIssue.Create(IssueCategory.BuildStep, step.name)
+                yield return ProjectIssue.Create(IssueCategory.BuildStep, step.name)
                     .WithCustomProperties(new object[(int)BuildReportStepProperty.Num]
                     {
                         Formatting.FormatBuildTime(step.duration),
                         step.name
                     })
                     .WithDepth(depth)
-                    .WithSeverity(Severity.Info));
+                    .WithSeverity(Severity.Info);
 
                 foreach (var message in step.messages)
                 {
                     var logMessage = message.content;
                     var description = new StringReader(logMessage).ReadLine(); // only take first line
-                    var issue = ProjectIssue.Create(IssueCategory.BuildStep, description)
+                    yield return ProjectIssue.Create(IssueCategory.BuildStep, description)
                         .WithCustomProperties(new object[(int)BuildReportStepProperty.Num]
                         {
                             0,
@@ -215,12 +164,11 @@ namespace Unity.ProjectAuditor.Editor.Modules
                         })
                         .WithDepth(depth + 1)
                         .WithSeverity(Diagnostic.Utils.LogTypeToSeverity(message.type));
-                    issues.Add(issue);
                 }
             }
         }
 
-        void AnalyzePackedAssets(BuildReport buildReport, IList<ProjectIssue> issues)
+        IEnumerable<ProjectIssue> AnalyzePackedAssets(BuildReport buildReport)
         {
             foreach (var packedAsset in buildReport.packedAssets)
             {
@@ -229,14 +177,10 @@ namespace Unity.ProjectAuditor.Editor.Modules
                 {
                     // sourceAssetPath might contain '|' which is invalid. This is due to compressed texture format names in the asset name such as DXT1|BC1
                     var assetPath = PathUtils.ReplaceInvalidChars(content.sourceAssetPath);
-
-                    // handle special case of Built-in assets
-                    if (assetPath.StartsWith("Built-in") && assetPath.Contains(":"))
-                        assetPath = assetPath.Substring(0, assetPath.IndexOf(':'));
-
-                    var description = string.IsNullOrEmpty(assetPath) ? k_Unknown : Path.GetFileNameWithoutExtension(assetPath);
                     var assetImporter = AssetImporter.GetAtPath(assetPath);
-                    var issue = ProjectIssue.Create(IssueCategory.BuildFile, description)
+                    var description = string.IsNullOrEmpty(assetPath) ? k_Unknown : Path.GetFileNameWithoutExtension(assetPath);
+
+                    yield return ProjectIssue.Create(IssueCategory.BuildFile, description)
                         .WithLocation(assetPath)
                         .WithCustomProperties(new object[(int)BuildReportFileProperty.Num]
                         {
@@ -245,16 +189,14 @@ namespace Unity.ProjectAuditor.Editor.Modules
                             content.packedSize,
                             packedAsset.shortPath
                         });
-                    issues.Add(issue);
                 }
             }
         }
 
-        void NewMetaData(string key, object value, IList<ProjectIssue> issues)
+        ProjectIssue NewMetaData(string key, object value)
         {
-            var issue = ProjectIssue.Create(IssueCategory.BuildSummary, key)
+            return ProjectIssue.Create(IssueCategory.BuildSummary, key)
                 .WithCustomProperties(new object[(int)BuildReportMetaData.Num] { value });
-            issues.Add(issue);
         }
 
 #endif