+++ Editor/Modules/ICodeModuleInstructionAnalyzer.cs
@@ -6,11 +6,9 @@ using Unity.ProjectAuditor.Editor.Core;
 
 namespace Unity.ProjectAuditor.Editor.Modules
 {
-    public interface ICodeModuleInstructionAnalyzer
+    public interface ICodeModuleInstructionAnalyzer : IModuleAnalyzer
     {
         IReadOnlyCollection<OpCode> opCodes { get; }
-
-        void Initialize(ProjectAuditorModule module);
 
         IssueBuilder Analyze(MethodDefinition methodDefinition, Instruction inst);
     }