+++ Tests/Editor/PathUtilTests.cs
@@ -10,6 +10,9 @@ namespace Unity.ProjectAuditor.EditorTests
         {
             Assert.AreEqual("sactx-0-2048x2048-DXT5_BC3-MenusAtlas-e85506f4",
                 PathUtils.ReplaceInvalidChars("sactx-0-2048x2048-DXT5|BC3-MenusAtlas-e85506f4"));
+
+            Assert.AreEqual("Built-in Texture2D sactx-0-256x256-ETC2-client_rays_1_spriteAtlas-0165171c",
+                PathUtils.ReplaceInvalidChars("Built-in Texture2D: sactx-0-256x256-ETC2-client_rays_1_spriteAtlas-0165171c"));
         }
     }
 }